requirements.txt
discord.py
sympy
discord-components